Patches inconsistency between CNES/ESA API: Duration is defined as integer in MAL/Java specification, but it may contain a fractional component of seconds in MAL specification.

+++ malbinary/src/main/java/fr/cnes/maljoram/encoding/DurationDecoder.java
@@ -25,6 +25,6 @@ package fr.cnes.maljoram.encoding;
 
 public interface DurationDecoder {
   
-  int decode(Decoder decoder) throws Exception;
+  double decode(Decoder decoder) throws Exception;
 
 }
